requirements.txt
flask==2.0.1
flask-cors==3.0.10
pillow==9.0.0
torch==1.9.0
torchvision==0.10.0
numpy==1.21.2
matplotlib==3.4.3
sympy==1.9
segmentation-models-pytorch==0.2.1 